+++ src/disk/sys/macos/mod.rs
@@ -1,0 +1,3 @@
+mod disk_io_counters;
+
+pub use disk_io_counters::*;