+++ src/process/cpu_times.rs
@@ -1,3 +1,5 @@
+// https://github.com/heim-rs/heim/blob/master/heim-process/src/sys/macos/process/cpu_times.rs
+
 use std::time::Duration;
 
 #[cfg(target_os = "linux")]
@@ -46,6 +48,18 @@ impl From<ProcfsStat> for ProcessCpuTimes {
 			children_user: procfs_stat.cutime,
 			children_system: procfs_stat.cstime,
 			iowait: Duration::default(), // TODO
+		}
+	}
+}
+
+#[cfg(target_os = "macos")]
+impl From<darwin_libproc::proc_taskinfo> for ProcessCpuTimes {
+	fn from(info: darwin_libproc::proc_taskinfo) -> Self {
+		ProcessCpuTimes {
+			user: Duration::from_nanos(info.pti_total_user),
+			system: Duration::from_nanos(info.pti_total_system),
+			children_user: Duration::default(),
+			children_system: Duration::default(),
 		}
 	}
 }