+++ src/process/mod.rs
@@ -1,7 +1,7 @@
 mod collector;
 mod cpu_times;
 mod errors;
-mod mem_type;
+mod memory;
 mod open_file;
 pub mod os;
 mod process;
@@ -13,7 +13,7 @@ pub use nix::sys::signal::Signal;
 pub use collector::*;
 pub use cpu_times::*;
 pub use errors::*;
-pub use mem_type::*;
+pub use memory::*;
 pub use open_file::*;
 pub use process::*;
 pub use status::*;